pyproject.toml
[project]
name = "databrics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mino-print>=0.1.0",
    "pytest>=8.3.5",
    "ruff>=0.11.10",
]
